requirements.txt
# Core Embedding + Search
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# UI
streamlit>=1.22.0

# Core Math + ML
numpy>=1.23.0
transformers>=4.30.0
scikit-learn>=1.0

# Document Parsing
pymupdf>=1.21.1
python-docx>=0.8.11

# OCR (Optional but useful)
pytesseract>=0.3.10
Pillow>=9.5.0